Move folder interfaces to models
